+++ docs/changelog.md
@@ -5,6 +5,7 @@
 - Adds support for software system/container instances in multiple deployment groups.
 - Fixes an issue where internal software systems/people are not included by the DSL formatter when the enterprise is not set.
 - The DSL formatter now removes empty tokens at the end of lines.
+- Adds support for recognising escaped newlines (\n) in tokens.
 
 ## 1.13.0 (3rd September 2021)